pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Marcio Martinez <marcioedumartinez@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
fastapi = "^0.111.0"
opentelemetry-distro = "^0.45b0"
opentelemetry-exporter-otlp = "^1.24.0"
opentelemetry-instrumentation = "^0.45b0"
opentelemetry-instrumentation-asgi = "^0.45b0"
opentelemetry-instrumentation-fastapi = "^0.45b0"


[tool.poetry.group.dev.dependencies]
poethepoet = "^0.26.1"

[tool.poe]
envfile = ".env"

[tool.poe.tasks]
serve = 'opentelemetry-instrument --metrics_exporter console uvicorn app.app:app --port 8000 --host 0.0.0.0'
dev = 'opentelemetry-instrument --metrics_exporter console,otlp uvicorn app.app:app --port 8000 --host 0.0.0.0 --reload'
export = 'poetry export -f requirements.txt --output requirements.txt'
up = 'docker compose -f docker/docker-compose.yml up -d --build'
down = 'docker compose -f docker/docker-compose.yml down'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
